+++ .github/workflows/dev.yml
@@ -8,6 +8,8 @@ env:
   REACT_APP_API_KEY: ${{ secrets.REACT_APP_API_KEY_DEV }}
   REACT_APP_ENDPOINT_URL: ${{ secrets.REACT_APP_ENDPOINT_URL_DEV }}
   REACT_APP_CORS_ORIGIN: ${{ secrets.REACT_APP_CORS_ORIGIN_DEV }}
+  REACT_APP_FIREBASE_VAPID_KEY: ${{ secrets.REACT_APP_FIREBASE_VAPID_KEY_DEV }}
+  REACT_APP_FIREBASE_CONFIG: ${{ secrets.REACT_APP_FIREBASE_CONFIG_DEV }}
 
 on:
   push:
