+++ .github/workflows/prod.yml
@@ -7,6 +7,8 @@ env:
   REACT_APP_RECAPTCHA_SITE_KEY: ${{ secrets.REACT_APP_RECAPTCHA_SITE_KEY }}
   REACT_APP_API_KEY: ${{ secrets.REACT_APP_API_KEY_PROD }}
   REACT_APP_ENDPOINT_URL: ${{ secrets.REACT_APP_ENDPOINT_URL_PROD }}
+    REACT_APP_FIREBASE_VAPID_KEY: ${{ secrets.REACT_APP_FIREBASE_VAPID_KEY_PROD }}
+  REACT_APP_FIREBASE_CONFIG: ${{ secrets.REACT_APP_FIREBASE_CONFIG_PROD }}
 
 on:
   push:
