+++ package.json
@@ -14,6 +14,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "@types/async": "^3.2.16",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.16",
@@ -46,6 +47,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test --ci --detectOpenHandles --color --runInBand --watchAll --forceExit --coverage",
     "test:ci": "react-scripts test --ci --detectOpenHandles --color --runInBand --env=jsdom --watchAll=false",
+    "emulate:hosting": "react-scripts build && firebase serve --only hosting:test -p 3000",
     "eject": "react-scripts eject",
     "lint:ts": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
     "lint:css": "stylelint --allow-empty-input \"src/**/*.{css,scss}\"",
