+++ src/components/NavBar/index.scss
@@ -241,6 +241,7 @@
   font-size: 28px;
   color: white;
   margin-left: -0.75em;
+
   @media screen and (max-width: 400px) {
     display: none;
   }
