+++ src/views/Portal/index.scss
@@ -3,7 +3,7 @@
   flex-direction: column;
   padding: 0;
   flex-basis: 100%;
-    justify-content: center;
+  justify-content: center;
   align-items: center;
 }
