+++ webpack-config.js
@@ -32,6 +32,8 @@ module.exports = {
         REACT_APP_API_KEY: JSON.stringify(process.env.REACT_APP_API_KEY),
         REACT_APP_ENDPOINT_URL: JSON.stringify(process.env.REACT_APP_ENDPOINT_URL),
         REACT_APP_CORS_ORIGIN: JSON.stringify(process.env.REACT_APP_CORS_ORIGIN),
+        REACT_APP_FIREBASE_VAPID_KEY: JSON.stringify(REACT_APP_FIREBASE_VAPID_KEY),
+        REACT_APP_FIREBASE_CONFIG: JSON.stringify(REACT_APP_FIREBASE_CONFIG),
       },
     }),
   ],
